+++ lib/Model/CreateCommentPublic200Response.php
@@ -66,8 +66,8 @@ class CreateCommentPublic200Response implements ModelInterface, ArrayAccess, \Js
         'reason' => 'string',
         'code' => 'string',
         'secondary_code' => 'string',
-        'banned_until' => 'float',
-        'max_character_length' => 'float',
+        'banned_until' => 'int',
+        'max_character_length' => 'int',
         'translated_error' => 'string',
         'custom_config' => '\FastComments\Client\Model\CustomConfigParameters'
     ];
@@ -88,8 +88,8 @@ class CreateCommentPublic200Response implements ModelInterface, ArrayAccess, \Js
         'reason' => null,
         'code' => null,
         'secondary_code' => null,
-        'banned_until' => 'double',
-        'max_character_length' => 'double',
+        'banned_until' => 'int64',
+        'max_character_length' => 'int32',
         'translated_error' => null,
         'custom_config' => null
     ];
@@ -608,7 +608,7 @@ class CreateCommentPublic200Response implements ModelInterface, ArrayAccess, \Js
     /**
      * Gets banned_until
      *
-     * @return float|null
+     * @return int|null
      */
     public function getBannedUntil()
     {
@@ -618,7 +618,7 @@ class CreateCommentPublic200Response implements ModelInterface, ArrayAccess, \Js
     /**
      * Sets banned_until
      *
-     * @param float|null $banned_until banned_until
+     * @param int|null $banned_until banned_until
      *
      * @return self
      */
@@ -635,7 +635,7 @@ class CreateCommentPublic200Response implements ModelInterface, ArrayAccess, \Js
     /**
      * Gets max_character_length
      *
-     * @return float|null
+     * @return int|null
      */
     public function getMaxCharacterLength()
     {
@@ -645,7 +645,7 @@ class CreateCommentPublic200Response implements ModelInterface, ArrayAccess, \Js
     /**
      * Sets max_character_length
      *
-     * @param float|null $max_character_length max_character_length
+     * @param int|null $max_character_length max_character_length
      *
      * @return self
      */
